requirements.txt
fastapi
uvicorn
pydantic
llama-cpp-python
motor
sentence-transformers
faiss-cpu
numpy
asyncio
pymongo
aiohttp
httpx
